+++ lib/action/context.rb
@@ -1,28 +1,38 @@
 # frozen_string_literal: true
 
-# NOTE: This is a temporary file to be removed when we have a better way to handle context.
-# rubocop:disable Style/OpenStructUse, Style/CaseEquality
-require "ostruct"
-
 module Action
-  class Context < OpenStruct
-    def self.build(context = {})
-      self === context ? context : new(context)
+  class Context
+    attr_accessor :provided_data, :exposed_data
+
+    def initialize(**provided_data)
+      @provided_data = provided_data
+      @exposed_data = {}
+
+      # Framework-managed fields
+      @failure = false
+      @exception = nil
+      @error_from_user = nil
+      @error_prefix = nil
+      @elapsed_time = nil
     end
 
-    def success?
-      !failure?
+    def fail!(message = nil)
+      @error_from_user = message if message.present?
+      raise Action::Failure, message
     end
 
-    def failure?
-      @failure || false
-    end
+    # INTERNAL: base for further filtering (for logging) or providing user with usage hints
+    def __combined_data = @provided_data.merge(@exposed_data)
 
-    def fail!(context = {})
-      context.each { |key, value| self[key.to_sym] = value }
-      @failure = true
-      raise Action::Failure, self
-    end
+    # Framework state methods
+    def ok? = !@failure
+    def failed? = @failure || false
+
+    # Framework field accessors
+    attr_accessor :exception, :error_from_user, :error_prefix, :elapsed_time
+
+    # Internal failure state setter (for framework use)
+    attr_writer :failure
+    private :failure=
   end
 end
-# rubocop:enable Style/OpenStructUse, Style/CaseEquality